# harmony-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Harmony API",
    "description": "REST API for Harmony.\n\n**Rate Limiting:** All endpoints are subject to rate limiting. When the limit is exceeded, the API responds with `429 Too Many Requests` and a `Retry-After` header indicating when the client may retry.",
    "license": {
      "name": "AGPL-3.0"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/v1/channels/{id}/voice/leave": {
      "post": {
        "tags": [
          "Voice"
        ],
        "summary": "Leave a voice channel. The user must be in the specified channel.",
        "description": "# Errors\n- 401 if not authenticated.\n- 409 if the user is not in the specified channel.\n- 503 if voice is not configured (`LiveKit` disabled).",
        "operationId": "leave_voice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Voice channel ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ChannelId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Left voice channel"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Not in the specified channel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Voice not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/v1/servers/{id}/leave": {
      "post": {
        "tags": [
          "Members"
        ],
        "summary": "Leave a server voluntarily. The owner cannot leave (must transfer ownership first).",
        "description": "# Errors\nReturns `ApiError` on validation failure or repository error.",
        "operationId": "leave_server",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ServerId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Left the server"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Owner cannot leave",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Server not found or not a member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/servers/{id}/members": {
      "get": {
        "tags": [
          "Members"
        ],
        "summary": "List members of a server with cursor-based pagination.",
        "description": "Use `before` (ISO 8601) to paginate backward. Default limit is 50, max is 100.\n\n# Errors\nReturns `ApiError` if the cursor is invalid or a repository error occurs.",
        "operationId": "list_members",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ServerId"
            }
          },
          {
            "name": "before",
            "in": "query",
            "description": "ISO 8601 timestamp cursor -- fetch members who joined before this time.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of members to return (1-100, default 50).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid cursor or limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Server not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Members"
        ],
        "summary": "Join a server via an invite code.",
        "description": "Validates the invite, checks that the user is not already a member,\nand adds them to the server.\n\n# Errors\nReturns `ApiError` on invalid invite, expired invite, or conflict.",
        "operationId": "join_server",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ServerId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinServerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Joined successfully"
          },
          "400": {
            "description": "Invalid or expired invite",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Already a member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/servers/{id}/members/{user_id}": {
      "delete": {
        "tags": [
          "Members"
        ],
        "summary": "Kick a member from a server. Requires moderator+ role with hierarchy enforcement.",
        "description": "# Errors\nReturns `ApiError` on authorization failure or repository error.",
        "operationId": "kick_member",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ServerId"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "description": "User ID to kick",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Member kicked"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient role or hierarchy violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Server or member not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/servers/{id}/members/{user_id}/role": {
      "patch": {
        "tags": [
          "Members"
        ],
        "summary": "Assign a role to a server member. Requires admin+ role with hierarchy enforcement.",
        "description": "# Errors\nReturns `ApiError` on validation failure, authorization failure, or repository error.",
        "operationId": "assign_role",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ServerId"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "description": "Target user ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Role assigned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid role or self-assignment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient role or hierarchy violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Server or member not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/v1/servers/{id}/transfer-ownership": {
      "post": {
        "tags": [
          "Members"
        ],
        "summary": "Transfer server ownership. Only the current owner can do this.",
        "description": "# Errors\nReturns `ApiError` on authorization failure or repository error.",
        "operationId": "transfer_ownership",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ServerId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferOwnershipRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ownership transferred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Cannot transfer to self",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Not server owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Server or new owner not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignRoleRequest": {
        "type": "object",
        "description": "Request body for assigning a role to a server member.",
        "required": [
          "role"
        ],
        "properties": {
          "role": {
            "$ref": "#/components/schemas/Role",
            "description": "The role to assign (admin, moderator, member). Use transfer-ownership for owner."
          }
        },
        "additionalProperties": false
      },
...
      "ChannelId": {
        "type": "string",
        "format": "uuid",
        "description": "Unique identifier for a channel within a server.",
        "example": "770e8400-e29b-41d4-a716-446655440002"
      },
...
      "JoinServerRequest": {
        "type": "object",
        "description": "Request body for joining a server via invite code.",
        "required": [
          "inviteCode"
        ],
        "properties": {
          "inviteCode": {
            "type": "string",
            "description": "The invite code to use."
          }
        },
        "additionalProperties": false
      },
...
      "MemberListResponse": {
        "type": "object",
        "description": "Envelope for a list of server members with cursor pagination (ADR-036).",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberResponse"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Cursor for the next page. `None` if this is the last page."
          }
        }
      },
      "MemberResponse": {
        "type": "object",
        "description": "Server member response returned to API consumers.",
        "required": [
          "userId",
          "username",
          "role",
          "joinedAt"
        ],
        "properties": {
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          },
          "nickname": {
            "type": [
              "string",
              "null"
            ]
          },
          "role": {
            "type": "string"
          },
          "userId": {
            "$ref": "#/components/schemas/UserId"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 9457 Problem Details response.\n\nAll API errors are returned in this standardized format.",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "upgrade_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "WHY: Drives upsell when a plan limit is exceeded. The frontend renders\nan \"Upgrade\" CTA using this URL. Only present on `LimitExceeded` errors."
          }
        }
      },
...
      "ServerId": {
        "type": "string",
        "format": "uuid",
        "description": "Unique identifier for a server (guild).",
        "example": "660e8400-e29b-41d4-a716-446655440001"
      },
...
      "UserId": {
        "type": "string",
        "format": "uuid",
        "description": "Unique identifier for a user (Supabase auth.users UUID).",
        "example": "550e8400-e29b-41d4-a716-446655440000"
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT (mobile/API clients)"
      }
    }
  },
...
}